Create Substitution Cipher
Very happy with this one. Took what I learned from the caesar cipher algorithm to create a substitution algorithm. User inputs a 26 long string of letters as the command argument, which runs through some checks. Then essentially with C = (P + K)%26, sets the resulting ciphText character to the corresponding key character. 

First it makes the key array[0] the first character of the key, key[1] the second character and so on, up until key[25], with some ascii conversion before.
Then it finds the ascii value of the plaintext, lower and uppercase dependant, so a would = 0, b = 1, and so on. 
Then basically makes it so plain[letter] value = the index of the key, resulting in the cipher.